parallax implementation with core/cover, some js cleanup, rebuilt theme
