+++ src/scss/components/_learn-more.scss
@@ -1,7 +1,6 @@
 $spacing: get-spacing();
 
 .learn-more-message {
-	background-color: color(green-dark) !important;
 	margin-top: 0;
 	margin-bottom: 0;
 	> h2 {
