+++ src/scss/components/_page__nav.scss
@@ -31,8 +31,8 @@
 		padding: 0;
 		gap: 0;
 		box-shadow: 0 0.5rem 1rem rgba(0, 0, 0, 0.3);
-
 		top: 0;
+
 		.admin-bar & {
 			top: 32px;
 		}
