+++ src/scss/index.scss
@@ -27,3 +27,4 @@
 @import 'components/button';
 @import 'components/campaigns';
 @import 'components/our-goals';
+@import 'components/parallax';
